Support version specific behaviour
In version 0.9.1 and the 0.10 dev branch the layout of functions is different, one using a table and one using divs. In 0.9.1, placing before the parent element causes all the anchors to end up before the tableso we must place them as the first child of the same parent. In 0.10, placing as the first child of the same parent affects styling and highlighting in dash so we place it before the parent.